Move set handlers and splices to own modules
In preparation for limiting them to logged-in users only.
